requirements.txt
pygsheets==2.0.2
discord.py==1.2.5
